requirements.txt
flask
firebase-admin
google-genai
python-dotenv
cloudinary
google-api-python-client
google-auth
google-auth-httplib2